# swagger.json
{"openapi":"3.0.0","info":{"title":"API 文档","version":"1.0.0","description":"基于 Express 的通用 API"},"paths":{"/login/sign":{"post":{"tags":["/login"],"summary":"登陆","externalDocs":{"description":"登陆","url":"/login/sign"},"parameters":[],"requestBody":{"description":"用户信息","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","description":"用户名","required":["string"]},"password":{"type":"string","description":"密码","required":["string"]},"code":{"type":"string","description":"验证码","required":["string"]}}}}},"required":true},"responses":{"200":{"description":"成功"}}}},"/login/out":{"post":{"tags":["/login"],"summary":"登出","externalDocs":{"description":"登出","url":"/login/out"},"parameters":[],"requestBody":{"description":"用户信息","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string","description":"用户名ID","required":["string"]}}}}},"required":true},"responses":{"200":{"description":"成功"}}}},"/users/":{"get":{"tags":["/users"],"summary":"获取用户列表","externalDocs":{"description":"获取用户列表","url":"/users/"},"parameters":[{"in":"query","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"query","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"查询成功"}}}},"/users/get/{id}":{"get":{"tags":["/users"],"summary":"获取用户详情","externalDocs":{"description":"获取用户详情","url":"/users/get/{id}"},"parameters":[{"name":"id","in":"path","required":true,"description":"用户id","schema":{"type":"string"}}],"requestBody":{},"responses":{"200":{"description":"查询成功"}}}},"/users/create":{"post":{"tags":["/users"],"summary":"创建用户","externalDocs":{"description":"创建用户","url":"/users/create"},"parameters":[],"requestBody":{"description":"用户信息","content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string","description":"用户名","required":[]},"password":{"type":"string","description":"密码","required":[]}}}}},"required":true},"responses":{"200":{"description":"创建成功"}}}},"/users/update":{"post":{"tags":["/users"],"summary":"更新用户","externalDocs":{"description":"更新用户","url":"/users/update"},"parameters":[],"requestBody":{"description":"用户信息","content":{"application/json":{"schema":{"type":"object","properties":{"userId":{"type":"string","description":"用户id","required":["string"]},"username":{"type":"string","description":"用户名","required":["string"]},"sex":{"type":"string","description":"性别","required":[]},"email":{"type":"string","description":"邮箱","required":[]},"phone":{"type":"string","description":"手机号","required":[]}}}}},"required":true},"responses":{"200":{"description":"更新成功"},"500":{"description":"更新失败"}}}},"/upload/file":{"post":{"tags":["/upload"],"summary":"上传文件","externalDocs":{"description":"上传文件","url":"/upload/file"},"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"上传成功"},"500":{"description":"文件不能为空"}}}},"/upload/picture":{"post":{"tags":["/upload"],"summary":"上传照片","externalDocs":{"description":"上传照片","url":"/upload/picture"},"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"上传成功"},"500":{"description":"文件不能为空"}}}},"/upload/avatar":{"post":{"tags":["/upload"],"summary":"上头像片","externalDocs":{"description":"上头像片","url":"/upload/avatar"},"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"description":"上传成功"},"500":{"description":"文件不能为空"}}}},"/logs/login":{"get":{"tags":["/logs"],"summary":"登录日志","externalDocs":{"description":"登录日志","url":"/logs/login"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"获取登录日志成功","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/logs/users":{"get":{"tags":["/logs"],"summary":"获取用户日志","externalDocs":{"description":"获取用户日志","url":"/logs/users"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"获取用户日志成功","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/logs/upload":{"get":{"tags":["/logs"],"summary":"文件上传日志","externalDocs":{"description":"文件上传日志","url":"/logs/upload"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"获取文件上传日志成功","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/logs/stat":{"get":{"tags":["/logs"],"summary":"统计日志","externalDocs":{"description":"统计日志","url":"/logs/stat"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"获取统计日志成功","content":{"application/json":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/stat/login":{"get":{"tags":["stat"],"summary":"登录统计","externalDocs":{"description":"登录统计","url":"/stat/login"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"success","content":{"application/json":{"schema":{}}}},"500":{"description":"获取登录统计失败"}}}},"/stat/users":{"get":{"tags":["stat"],"summary":"用户统计","externalDocs":{"description":"用户统计","url":"/stat/users"},"parameters":[{"in":"path","name":"page","required":true,"description":"页码","schema":{"type":"integer"}},{"in":"path","name":"limit","required":true,"description":"每页数量","schema":{"type":"integer"}}],"requestBody":{},"responses":{"200":{"description":"success","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"number","description":"用户ID"},"name":{"type":"string","description":"用户名"},"age":{"type":"number","description":"年龄"}}}},"total":{"type":"number","description":"总记录数"},"page":{"type":"number","description":"当前页码"},"pageSize":{"type":"number"}}}}}},"500":{"description":"获取用户统计失败"}}}}},"components":{},"tags":[]}